requirements.txt
openai==1.40.8
streamlit==1.37.1
python-dotenv==1.0.1
gradio==4.44.1
ollama==0.3.3